pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prompt-generator-agent"
version = "0.1.0"
description = "A modular AI agent for generating, reviewing, and testing prompts"
authors = [{name = "Your Name"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "litellm>=1.0.0",
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    "streamlit>=1.28.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
prompt-generator-cli = "src.cli:main"
prompt-generator-ui = "src.ui:main"

[tool.setuptools]
packages = ["src"]